+++ engine/platform/unix/unix_shared.cpp
@@ -102,7 +102,7 @@ qboolean Sys_RandomBytes( byte *string, int len )
 
 	setvbuf( fp, NULL, _IONBF, 0 ); // don't buffer reads from /dev/urandom
 
-	if ( fread( string, sizeof( byte ), len, fp ) != len ) {
+	if ( fread( string, sizeof( byte ), len, fp ) != (size_t)len ) {
 		fclose( fp );
 		return qfalse;
 	}
@@ -562,7 +562,7 @@ uint64_t Sys_GetAffinityMask( void )
 	if ( sched_getaffinity( getpid(), sizeof( cpu_set ), &cpu_set ) == 0 ) {
 		uint64_t mask = 0;
 		int cpu;
-		for ( cpu = 0; cpu < sizeof( mask ) * 8; cpu++ ) {
+		for ( cpu = 0; (size_t)cpu < sizeof( mask ) * 8; cpu++ ) {
 			if ( CPU_ISSET( cpu, &cpu_set ) ) {
 				mask |= (1ULL << cpu);
 			}
@@ -585,7 +585,7 @@ qboolean Sys_SetAffinityMask( const uint64_t mask )
 	int cpu;
 
 	CPU_ZERO( &cpu_set );
-	for ( cpu = 0; cpu < sizeof( mask ) * 8; cpu++ ) {
+	for ( cpu = 0; (size_t)cpu < sizeof( mask ) * 8; cpu++ ) {
 		if ( mask & (1ULL << cpu) ) {
 			CPU_SET( cpu, &cpu_set );
 		}
